rs_download_API/pyproject.toml
[tool.poetry]
name = " rs-api "
version = "0.1.0"
description = "This project is an api tool to download rs products"
authors = ["Antonios Koutroumpas <antonkout@gmail.com>"]
license = "Copyright 2023 AntonKout, all rights reserved."

[tool.poetry.dependencies]
python = "^3.8"
dash = "2.10.2"
google-cloud-bigquery = { version = "2.20.0", python = ">=3.6,<3.10" }
google-auth = "1.31.0"
pandas = "1.3.4"
sentinelsat = "1.0.1"
folium = "0.12.1"
geopandas = "0.9.0"
shapely = "1.7.1"
landsatxplore = "0.12.1"
pyproj = "2.6.1"

[tool.pytest.ini_options]
filterwarnings = [
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"